Fix type mismatches, improve error handling, and add specs/plans
- Fix BaselineMetrics.AvgDurationMs int64→float64 to match template functions
- Remove unnecessary div wrapper in retro template duration comparison
- Show actual elapsed duration in timeout errors instead of configured timeout
- Fall through on learnings filter error instead of blocking learning placement
- Add methodology config section (atdd/tdd)
- Add specs and plans for debug-command, epic-scoped-execution, parallel-post-success
- Update learnings, state, and bead tracker

+++ internal/retro/experiment.go
@@ -10,7 +10,7 @@ import (
 // BaselineMetrics holds efficiency metrics captured at the start of an experiment
 type BaselineMetrics struct {
 	AvgCostPerBead  float64 `json:"avg_cost_per_bead"`
-	AvgDurationMs   int64   `json:"avg_duration_ms"`
+	AvgDurationMs   float64 `json:"avg_duration_ms"`
 	AvgInputTokens  float64 `json:"avg_input_tokens"`
 	AvgOutputTokens float64 `json:"avg_output_tokens"`
 	FailureRate     float64 `json:"failure_rate"`
@@ -127,7 +127,7 @@ func ComputeBaselineMetrics(
 
 	return BaselineMetrics{
 		AvgCostPerBead:  avgCost,
-		AvgDurationMs:   avgDuration.Milliseconds(),
+		AvgDurationMs:   float64(avgDuration.Milliseconds()),
 		AvgInputTokens:  avgInput,
 		AvgOutputTokens: avgOutput,
 		FailureRate:     stats.FailureRate(),
